Update streamlit_TS.py
For step 7, instead of asking them to attach, which 70% will miss, pls add a mandatory field in the first step itself to enter thier TS Tracker Link and include with the generated table in step 4 - if possible with a hyper link like <Employee Name> TS Tracker



As seen in my screenshot, the font appears too small on my macOS. Not sure what size is seen on Windows. Would it be possible to increase the font size?

3. Additional option to show the month's total and option to select the month.

4. When option to open TS Submission Status is clicked, that sheet is not opening for me since a couple of months. Not sure what is the issue. Have not blocked any popup I think..Chrome. (edited) 
5. Here, below the 'Timesheet generated successfully' message, one checkbox can be given for Have you pasted this week's timesheet data in your online tracker?
If not checked, don't allow them to send the email. I know this cannot ensure they have pasted it, but it will be a good reminder and most will do it.  (edited) 
6. Can we get the Vyoma logo at the top left, so that the screenshot can look like this.
7. The count starts from zero - in the first column. Should start from 1.
THen it continues beyond the Tasks..